% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shm_create_segment}
\alias{shm_create_segment}
\title{Create shared memory segment}
\usage{
shm_create_segment(segment, size)
}
\arguments{
\item{segment}{segment name}

\item{size}{initial size in bytes}
}
\description{
Create shared memory segment
}
