% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shm_get}
\alias{shm_get}
\title{Get object from shared memory}
\usage{
shm_get(segment, name)
}
\arguments{
\item{segment}{segment name}

\item{name}{object name}
}
\description{
Get object from shared memory
}
