% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shm_remove_segment}
\alias{shm_remove_segment}
\title{Remove shared memory segment}
\usage{
shm_remove_segment(segment)
}
\arguments{
\item{segment}{segment name}
}
\description{
Remove shared memory segment
}
