% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shm_set}
\alias{shm_set}
\title{Set object in shared memory}
\usage{
shm_set(x, segment, name)
}
\arguments{
\item{x}{an R object}

\item{segment}{a string}

\item{name}{a string}

\item{overwrite}{Overwrite segment if exists}
}
\description{
Set object in shared memory
}
